refactor(react/runtime): move `delayedRunOnMainThreadData` to a separate file
