+++ packages/react/runtime/src/lynx/tt.ts
@@ -21,8 +21,8 @@ import { CHILDREN } from '../renderToOpcodes/constants.js';
 import { __root } from '../root.js';
 import { backgroundSnapshotInstanceManager } from '../snapshot.js';
 import type { SerializedSnapshotInstance } from '../snapshot.js';
+import { delayedRunOnMainThreadData, takeDelayedRunOnMainThreadData } from '../worklet/delayedRunOnMainThreadData.js';
 import { destroyWorklet } from '../worklet/destroy.js';
-import { delayedRunOnMainThreadData, takeDelayedRunOnMainThreadData } from '../worklet/runOnMainThread.js';
 
 export { runWithForce };
